Require token for any API call after login.

+++ app/trades.py
@@ -6,6 +6,9 @@ from app.models import User, Trade
 
 @app.route('/trade/id/<int:id>')
 def get_trade_by_id(id: int) -> dict:
+    if not tokens.is_valid(request):
+        return 'Invalid or expired token.', 401
+    
     trade = Trade.query.get(id)
     if trade is not None:
         return trade.json()
@@ -14,6 +17,9 @@ def get_trade_by_id(id: int) -> dict:
 
 @app.route('/trade/user/<string:username>')
 def get_trades_by_username(username: str) -> list[dict]:
+    if not tokens.is_valid(request):
+        return 'Invalid or expired token.', 401
+
     user = get_user(username)
     trades = Trade.query.filter_by(user=user).all()
     return [t.json() for t in trades]
@@ -24,7 +30,7 @@ def get_user(username):
 
 @app.route('/trade', methods=['POST'])
 def trade() -> list[dict]:
-    if not token_is_valid(request):
+    if not tokens.is_valid(request):
         return 'Invalid or expired token.', 401
 
     new_trade = create_trade(request)
@@ -35,15 +41,6 @@ def trade() -> list[dict]:
     db.session.commit()
 
     return [new_trade.json(), cash_trade.json()], 201
-
-def token_is_valid(request: Request) -> bool:
-    token = get_token(request)
-    return tokens.is_valid(token) and not tokens.is_expired(token)
-
-def get_token(request: Request) -> str:
-    auth_header = request.headers.get('Authorization')
-    token = auth_header.split(' ')[1]
-    return token
 
 def create_trade(request: Request) -> Trade:
     trade_obj = request.json
@@ -67,6 +64,9 @@ def create_cash_transaction(trade: Trade) -> Trade:
 
 @app.route('/trade/id/<int:id>', methods=['DELETE'])
 def delete_trade(id: int) -> Response:
+    if not tokens.is_valid(request):
+        return 'Invalid or expired token.', 401
+
     trade = Trade.query.get(id)
     if trade is not None:
         db.session.delete(trade)
